Split build into multiple stages
The goal is to build the arm packages on the native cpu of the host then do the finall install inside of the target environment.
Right now building all three architectures on github takes around 5 hours. I'm hoping to get that down to under an hour with this patch set.
Currently this appears to build amd64 fine but cross compiling for arm is not working.